# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "SmartShop API",
        "contact": {},
        "license": {},
        "version": "v1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/ping": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Test Ping",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "We need ID!!",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Can not find ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/users": {
            "post": {
                "description": "创建用户",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名字段",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户昵称",
                        "name": "nickname",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "男性",
                        "name": "male",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "错误码",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/helpers.ErrorBody"
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "put": {
                "description": "更新用户",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名字段",
                        "name": "username",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "用户昵称",
                        "name": "nickname",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "男性",
                        "name": "male",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "错误码",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/helpers.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.ErrorBody": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer"
                },
                "local": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "male": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}